server/eezy_store/requirements.txt
asgiref==3.8.1
autopep8==2.3.2
Django==5.2
django-cors-headers==4.7.0
django-debug-toolbar==5.1.0
django-ipware==7.0.1
django-silk==5.3.2
django-structlog==9.1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gprof2dot==2025.4.14
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
pycodestyle==2.13.0
PyJWT==2.9.0
python-ipware==3.0.0
sqlparse==0.5.3
structlog==25.2.0
